apps/core/pyproject.toml
[project]
name = "cerise-core"
version = "0.1.0"
description = "AI-driven Live2D Virtual Character Core"
# readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "httpx>=0.28.1",
    "numpy>=2.4.1",
    "aiohttp>=3.13.3",
]

[project.optional-dependencies]
openai = [
    "openai>=2.15.0",
]
claude = [
    "anthropic>=0.76.0",
]
gemini = [
    "google-generativeai>=0.8.6",
]
all = [
    "openai>=2.15.0",
    "anthropic>=0.76.0",
    "google-generativeai>=0.8.6",
]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
    "mypy>=1.19.1",
]

[project.scripts]
cerise = "apps.core.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["abilities", "ai", "api", "character", "config", "infrastructure", "plugins"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
